requirements.txt
bcrypt==3.2.2
blueprint==3.4.2
cffi==1.15.0
click==8.0.4
colorama==0.4.4
DateTime==4.4
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.0
Jinja2==3.1.2
MarkupSafe==2.1.0
Pillow==9.1.1
platformdirs==2.5.1
pycparser==2.21
pytz==2022.1
SQLAlchemy==1.4.37
tabulate==0.8.9
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.0
